+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Aldryn Boilerplate Bootstrap3
 3.3.1
 =====
 - work on documentation
+- improve local testing
 
 
 3.3.0